restructured system
Use channels to distribute messages over N threads, where N is the number of agents. Use sync.Map to allow agents to be added while current agents are reading from the map (i.e. listening to their channels).

Also added a read write lock to the sendMessage function as well as a flag that indicates whether it is safe to send a message (will be true before channel closure and false after).